backend/JUDEMENT_RAG/requirements.txt
# Core dependencies for Legal Judgment RAG System
langchain==0.1.5
langchain-community==0.0.19
langchain-core==0.1.22
langchain-text-splitters==0.0.1

# PDF processing
pypdf==3.17.4
pdfplumber==0.10.3
unstructured==0.12.0
unstructured-inference==0.7.21

# Vector embeddings and storage
sentence-transformers==2.3.1
faiss-cpu==1.7.4

# LLM integration
transformers==4.37.2
huggingface-hub==0.20.3
accelerate==0.26.1
safetensors==0.4.2
torch==2.1.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
numpy==1.26.2
pandas==2.1.3

# Web interface (optional)
gradio==4.19.1